pyproject.toml
[tool.poetry]
name = "sports-prediction"
version = "0.1.0"
description = ""
authors = ["miguelm7 <miguelmore100@gmail.com>"]
readme = "README.md"
packages = [{include = "sports_prediction"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
